# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Renoly v2 API
  version: 0.1.0
  description: |
    Machine-readable contract for the clean-slate Renoly v2 API.
    The normative prose remains docs/api-spec.md. Employee operations always
    verify the path organization, active membership, role and resource tenant.
    Cross-tenant resources are returned as 404 to avoid existence disclosure.
  contact:
    name: Renoly engineering
servers:
  - url: /api/v2
    description: Same-origin application API
...
paths:
...
  /organizations/{orgId}/work-orders/{id}/actions/schedule:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/ResourceId"
      - $ref: "#/components/parameters/IfMatch"
    post:
      tags: [Work orders]
      operationId: scheduleWorkOrder
      summary: Schedule and assign a work order
      security: *employeeWriteSecurity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleWorkOrderRequest"
      responses:
        <<: *standardErrors
        "200":
          description: Scheduled work order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderMutationEnvelope"
  /organizations/{orgId}/work-orders/{id}/actions/transition:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/ResourceId"
      - $ref: "#/components/parameters/IfMatch"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      tags: [Work orders]
      operationId: transitionWorkOrder
      summary: Apply a server-decided work-order transition
      description: Client sends an action, never a target status.
      security: *employeeWriteSecurity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionWorkOrderRequest"
      responses:
        <<: *standardErrors
        "200":
          description: Transitioned work order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderMutationEnvelope"
  /organizations/{orgId}/work-orders/{id}/assignments:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Work orders]
      operationId: listWorkOrderAssignments
      summary: List assignments for a work order
      responses:
        <<: *standardErrors
        "200":
          description: Assignment list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentListEnvelope"
    post:
      tags: [Work orders]
      operationId: createWorkOrderAssignment
      summary: Assign an active organization member
      security: *employeeWriteSecurity
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        <<: *standardErrors
        "201":
          description: Assignment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentEnvelope"
...
components:
  securitySchemes:
    employeeSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        Logical HttpOnly, Secure, SameSite session cookie issued by the server-side
        Supabase Auth adapter. Provider access and refresh tokens are never exposed.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required for employee mutations in addition to the session cookie.
    lineSignature:
      type: apiKey
      in: header
      name: X-Line-Signature
      description: Base64 HMAC-SHA256 signature verified over the exact raw request bytes.
    workerBearer:
      type: http
      scheme: bearer
      bearerFormat: short-lived service JWT
      description: Audience-restricted workload identity; never accepted by employee routes.
    publicCapabilityBearer:
      type: http
      scheme: bearer
      bearerFormat: 256-bit opaque capability
      description: Raw public quote capability carried only in Authorization; never in the HTTP path or logs.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      description: Organization tenant boundary checked against the active membership.
      schema:
        $ref: "#/components/schemas/Uuid"
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the last observed positive lock version.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
        examples: ['"3"']
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Scoped to actor, organization, method and route template.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    AssignmentDuty:
      type: string
      enum: [lead, technician, helper, observer]
...
    ListMeta:
      type: object
      additionalProperties: false
      properties:
        nextCursor:
          type: [string, "null"]
          minLength: 16
          maxLength: 1024
        hasMore:
          type: boolean
      required: [nextCursor, hasMore]
...
    WorkOrderDetailAssignment:
      type: object
      additionalProperties: false
      description: Assignment roster entry as projected onto the work-order detail.
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        membershipId:
          $ref: "#/components/schemas/Uuid"
        memberName:
          type: [string, "null"]
        duty:
          type: string
        status:
          type: string
        assignedAt:
          type: [string, "null"]
        acceptedAt:
          type: [string, "null"]
        declinedAt:
          type: [string, "null"]
        checkedInAt:
          type: [string, "null"]
        completedAt:
          type: [string, "null"]
        cancelledAt:
          type: [string, "null"]
        declineReason:
          type: [string, "null"]
        lockVersion:
          type: integer
      required: [id, membershipId, memberName, duty, status, assignedAt, acceptedAt, declinedAt, checkedInAt, completedAt, cancelledAt, declineReason, lockVersion]
    WorkOrderDetailChecklistItem:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        label:
          type: string
        responseType:
          type: string
        isRequired:
          type: boolean
        evidenceRequired:
          type: boolean
        options:
          description: Opaque option payload; shape depends on responseType.
        response:
          description: Opaque recorded response; shape depends on responseType.
        completedAt:
          type: [string, "null"]
        completedByMembershipId:
          type: [string, "null"]
          format: uuid
        sortOrder:
          type: integer
      required: [id, label, responseType, isRequired, evidenceRequired, options, response, completedAt, completedByMembershipId, sortOrder]
    WorkOrderDetailChecklist:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
        status:
          type: string
        completedAt:
          type: [string, "null"]
        completedByMembershipId:
          type: [string, "null"]
          format: uuid
        lockVersion:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderDetailChecklistItem"
      required: [id, name, status, completedAt, completedByMembershipId, lockVersion, items]
    WorkOrderDetailPhoto:
      type: object
      additionalProperties: false
      description: >-
        Photo projection. storagePath is emitted empty to clients; photo bytes are
        read only through the authorized signed-URL route.
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        category:
          type: string
        status:
          type: string
        checklistItemId:
          type: [string, "null"]
          format: uuid
        storagePath:
          type: string
        mimeType:
          type: string
        byteSize:
          type: [integer, "null"]
        width:
          type: [integer, "null"]
        height:
          type: [integer, "null"]
        sha256:
          type: [string, "null"]
        caption:
          type: [string, "null"]
        capturedAt:
          type: [string, "null"]
        uploadedByMembershipId:
          type: [string, "null"]
          format: uuid
        readyAt:
          type: [string, "null"]
        lockVersion:
          type: integer
        createdAt:
          type: string
      required: [id, category, status, checklistItemId, storagePath, mimeType, byteSize, width, height, sha256, caption, capturedAt, uploadedByMembershipId, readyAt, lockVersion, createdAt]
...
    WorkOrderDetail:
      type: object
      additionalProperties: false
      description: >-
        Full work-order detail projection returned by create, get, schedule,
        transition and force-complete. Includes the nested assignment roster,
        checklist snapshots and photo projections.
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        organizationId:
          $ref: "#/components/schemas/Uuid"
        workOrderNo:
          type: string
        projectId:
          type: [string, "null"]
          format: uuid
        serviceRequestId:
          type: [string, "null"]
          format: uuid
        customerId:
          $ref: "#/components/schemas/Uuid"
        locationId:
          $ref: "#/components/schemas/Uuid"
        assetId:
          type: [string, "null"]
          format: uuid
        title:
          type: string
        description:
          type: [string, "null"]
        customerNotes:
          type: [string, "null"]
        technicianNotes:
          type: [string, "null"]
        internalNotes:
          type: [string, "null"]
        completionSummary:
          type: [string, "null"]
        priority:
          type: string
        status:
          type: string
        scheduledStartAt:
          type: [string, "null"]
        scheduledEndAt:
          type: [string, "null"]
        dispatchedAt:
          type: [string, "null"]
        enRouteAt:
          type: [string, "null"]
        onSiteAt:
          type: [string, "null"]
        pausedAt:
          type: [string, "null"]
        completedAt:
          type: [string, "null"]
        cancelledAt:
          type: [string, "null"]
        cancellationReason:
          type: [string, "null"]
        requiresCustomerSignoff:
          type: boolean
        customerSignedAt:
          type: [string, "null"]
        lockVersion:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderDetailAssignment"
        checklists:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderDetailChecklist"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderDetailPhoto"
        replayed:
          type: boolean
      required: [id, organizationId, workOrderNo, projectId, serviceRequestId, customerId, locationId, assetId, title, description, customerNotes, technicianNotes, internalNotes, completionSummary, priority, status, scheduledStartAt, scheduledEndAt, dispatchedAt, enRouteAt, onSiteAt, pausedAt, completedAt, cancelledAt, cancellationReason, requiresCustomerSignoff, customerSignedAt, lockVersion, createdAt, updatedAt, assignments, checklists, photos]
...
    ScheduleAssignmentInput:
      type: object
      additionalProperties: false
      properties:
        membershipId:
          $ref: "#/components/schemas/Uuid"
        duty:
          $ref: "#/components/schemas/AssignmentDuty"
      required: [membershipId]
    ScheduleWorkOrderRequest:
      type: object
      additionalProperties: false
      properties:
        scheduledStartAt:
          $ref: "#/components/schemas/Timestamp"
        scheduledEndAt:
          $ref: "#/components/schemas/Timestamp"
        occurredAt:
          $ref: "#/components/schemas/Timestamp"
        assignments:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/ScheduleAssignmentInput"
        conflictOverrideReason:
          type: [string, "null"]
          minLength: 1
          maxLength: 2000
      required: [scheduledStartAt, scheduledEndAt, occurredAt, assignments]
    TransitionWorkOrderRequest:
      oneOf:
        - type: object
          additionalProperties: false
          properties:
            action:
              type: string
              enum: [dispatch, enRoute, arrive, pause, resume]
            occurredAt:
              $ref: "#/components/schemas/Timestamp"
            overrideReason:
              type: [string, "null"]
              minLength: 1
              maxLength: 2000
          required: [action, occurredAt]
        - type: object
          additionalProperties: false
          properties:
            action:
              type: string
              enum: [complete]
            occurredAt:
              $ref: "#/components/schemas/Timestamp"
            completionSummary:
              type: string
              minLength: 1
              maxLength: 10000
            customerSignoffName:
              type: [string, "null"]
              minLength: 1
              maxLength: 120
            overrideReason:
              type: [string, "null"]
              minLength: 1
              maxLength: 2000
          required: [action, occurredAt, completionSummary, customerSignoffName]
        - type: object
          additionalProperties: false
          properties:
            action:
              type: string
              enum: [cancel]
            occurredAt:
              $ref: "#/components/schemas/Timestamp"
            reason:
              type: string
              minLength: 1
              maxLength: 2000
            overrideReason:
              type: [string, "null"]
              minLength: 1
              maxLength: 2000
          required: [action, occurredAt, reason]
        - type: object
          additionalProperties: false
          properties:
            action:
              type: string
              enum: [reopen]
            occurredAt:
              $ref: "#/components/schemas/Timestamp"
            reason:
              type: string
              minLength: 1
              maxLength: 2000
            overrideReason:
              type: [string, "null"]
              minLength: 1
              maxLength: 2000
          required: [action, occurredAt, reason]
      discriminator:
        propertyName: action
...
    MutationNotificationStatus:
      type: string
      description: >-
        Honest customer/technician notification status for M5 mutation responses.
        LINE delivery is deferred to M6, so scheduling, dispatch and completion
        never send a message and always report not_sent.
      enum: [not_sent]
    WorkOrderMutationNotification:
      type: object
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/MutationNotificationStatus"
        reason:
          type: string
          description: Machine-readable reason the notification was not sent.
          maxLength: 200
      required: [status, reason]
    WorkOrderMutationEnvelope:
      type: object
      additionalProperties: false
      description: >-
        Work-order mutation response. Carries the updated detail plus the honest
        notification status, so the UI never implies a customer message was sent.
      properties:
        data:
          $ref: "#/components/schemas/WorkOrderDetail"
        notification:
          $ref: "#/components/schemas/WorkOrderMutationNotification"
      required: [data, notification]
...
    Assignment:
      type: object
      additionalProperties: false
      description: Assignment resource returned by create, update, cancel and respond.
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        organizationId:
          $ref: "#/components/schemas/Uuid"
        workOrderId:
          $ref: "#/components/schemas/Uuid"
        membershipId:
          $ref: "#/components/schemas/Uuid"
        duty:
          type: string
        status:
          type: string
        assignedAt:
          type: [string, "null"]
        acceptedAt:
          type: [string, "null"]
        declinedAt:
          type: [string, "null"]
        checkedInAt:
          type: [string, "null"]
        completedAt:
          type: [string, "null"]
        cancelledAt:
          type: [string, "null"]
        declineReason:
          type: [string, "null"]
        lockVersion:
          type: integer
      required: [id, organizationId, workOrderId, membershipId, duty, status, assignedAt, acceptedAt, declinedAt, checkedInAt, completedAt, cancelledAt, declineReason, lockVersion]
    CreateAssignmentRequest:
      type: object
      additionalProperties: false
      properties:
        membershipId:
          $ref: "#/components/schemas/Uuid"
        duty:
          $ref: "#/components/schemas/AssignmentDuty"
      required: [membershipId, duty]
...
    AssignmentEnvelope:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: "#/components/schemas/Assignment"
      required: [data]
    AssignmentListEnvelope:
      type: object
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Assignment"
        meta:
          $ref: "#/components/schemas/ListMeta"
      required: [data, meta]
...